chore: move anproc to a diff package
and replaced result-putting from using manual "=" ones, into

using a different function named `putResult` that throws an

exception when it founds a communication with the same name

in the result.
